Clean up properly.
I was getting occasional segfaults on exit. It turns out this was
because the FluidSynth objects have to be destroyed in the right order,
and they weren't being. However, once I'd adjusted A:FS to keep refs in
the right places I found it was still happening, because the FluidSynth
objects were surviving until the final phase of global destruction and
getting destroyed in the wrong order anyway.

So, I had to go on a circular reference hunt. After a fair amount of
mucking about, I believe I've caught everything.