Add a cursor.
It moves left and right and up and down, but insertion and backspacing
still currently happens at the end. I think before I fix that I want to
change Music to being a linked list, but I'm not quite sure how to get
the refcounts right.

Ideally, a section of Music would be a doubly-linked list joined in a
loop, with the pointer in indicating where we start. This allows us to
go back and forward, and makes it easy to find the end and remove a
section and insert a section and so on, but this is clearly a circular
structure.

Obviously there needs to be a rule that the forward refs are strong and
the backrefs are weak, but there is still a loop all the way around. I
don't think it's possible (or, at least, not easy) to have a Moo
attribute that's optionally weak; probably the only sensible answer is
to say that while the first element has a weak ref back to the last, the
last does not have a ref forward to the first. This would probably end
up being annoying, but it looks like something is going to have to be
annoying somewhere.

The other options that I can see are to have both ends unlinked and
trust that I can find an end if I need to, or to have some sort of
separate 'list' object which tracks both ends and has a weakref from
every element.