Insert and delete at the cursor.
This means changing Document->music from an array to a linklist.
Although this cascades to quite a lot of changes, they are mostly
simplifications.

I've decided that the list in ->music will always have a plain Music
item at its head. This greatly simplifies all the list traversal logic,
because I don't have to keep dealing with the case of an empty list. It
also means that a Cursor is now always positioned on a Music.

The MIDI player has changed most. I used to carefully build and cache an
array of just the information we needed for playing; this was because I
was worried about efficiency. I've changed this to just play off the
real list of notes in the document; this means that editing while
playing will play the edited notes without getting confused. (Previously
it would keep on playing the old notes, and sometimes the highlighting
would get confused when a SV address was reused for a new note.) If you
delete the currently-playing note it stops, since it's lost its pointer
to the notes: this is a reasonable thing to do, I think.

While I'm here, stick in some types, just for fun.