Add dots.
This requires notes with more width than just the notehead. I think I
want to change this so that ->draw moves the current point (or the whole
CTM? since glyphs are always shown at the origin?) along by the advance
width. This will be easier than tracking widths all over the place. (I'm
not sure what I'll do if/when I support multiple voices on a stave; or,
actually, multiple staves which must remain horizontally in sync.)

Dotted durations which are not a multiple of a qhdsquaver will not play
correctly. I ought to compensate for this by starting the note and then
skipping forward to the next point where we are properly quantised, but
that's not entirely straightforward since we need to keep hold of the
current note to know which pitch to turn off.