Create new Gtk classes for radio groups.
The vanilla RadioAction class has a number of problems, from my point of
view. First, there is no explicit representation of 'a group of radio
actions', instead the properties one would expect to see on the group
(the current-value property and the changed signal) are visible on every
member individually. Second, it's impossible to change the currently-
selected member of the group without emitting activate signals on both
the previously and newly selected members. Third, if the user explicitly
reactivates the currently-selected member (by key, button or menu), the
activation is ignored and no signals are emitted.

This is all awkward for my use. I want to be able to tie the current
value of the group to an attribute on Cursor; this means the actions end
up being activated every time the Cursor property changes. This in turn
means I can't (say) remove dots when a given note length is chosen,
because I can't readily distinguish between explicit user actions and an
incidental change in the Cursor's length caused by moving from one note
to the next.

This commit introduces a pair of classes, RadioGroup and RadioMember,
which solve this problem. They are both GtkAction subclasses, and their
behaviour is as follows:

    - A RadioGroup represents a group of radio actions. A RadioMember
      represents an individual action in the group. It's not currently
      possible to represent RadioGroups directly in the interface;
      instead, the individual members are represented, and activating
      them causes the group to be activated.

    - Members have a 'value' property, which is a string (rather than
      GtkRadioAction's integer). Groups have 'current' and
      'current-value' properties; 'current' is the currently-selected
      member, and 'current-value' is its value. Changing the current
      member with ->set_current{,_value} emits notify::current{,-value}
      but does not emit any activate signals.

    - An explicit activation on a group member changes the group's
      current member, and then emits activate on the group.

Because both GtkCheckMenuItem and GtkToggleToolItem assume their Action
will be a GtkToggleAction, and make C casts based on that assumption, I
have also had to add proxy classes for RadioMember. These inherit from
GtkCheckMenuItem and GtkToggleToolItem, but in their Activatable methods
they call up to the methods for GtkMenuItem and ToolItem rather than
their direct parents. This involved adding to the Perl bindings; the
Activatable interface was entirely unbound, and there was no facility
for calling parent implementations of an interface method. I've set
things up so calling (say) $self->Gtk2::MenuItem::UPDATE will invoke
GtkMenuItem's interface method, as expected in Perl.