Tidy up to to_lily methods.
Add one for Clef. Rename Music::Start to Music::Voice, since this object
represents the whole voice, and give it a to_lily method which iterates
over the list and returns the result. Fix the lily chroma.

Add a ToLily action which pops up a box with the generated lily in.