Give AddDot a toolbar button.
Also disable it when we aren't on something that HasLength. The disabled
appearance is less that ideal: perhaps because my icons are mostly
black-and-white, it's not very different from the enabled appearance. I
ought to be able to fix this by subclassing Gtk2::Style, but that might
require a whole lot more XS.