Key signatures.
This required quite a lot of other changes...

The Cairo and font information which was previously passed into ->draw
is now held in a DrawCtx object. This object now provides the glyph
methods previously provided by HasGlyph, and also keeps track of the
current clef. This is necessary because the positioning of the sharps
and flats depends on the clef, not just the centre line.

In fact, the detailed positioning of the accidentals is deeply
idiosyncratic. For instance, in treble clef F# goes up, but in tenor
clef it goes down. In bass clef I would expect A# to go up, but instead
it goes down. And so on. (The flats make more sense, at least in the
ordinary clefs, but I think this is more by chance than anything.) In
order to get this right, KeySig includes specific positioning for each
clef, meaning that I now don't support more than treble-alto-tenor-bass.

Since HasCentre now also provides a clef name, I've renamed it to
HasClef. Ottava items without an explicit clef will need to link back to
the previous real clef, in order to find which symbol they should use.

Sharpen and Flatten when in edit mode and on a key signature change the
key. There is currently no other way to insert or edit a key signature,
because I haven't written the dialog yet. Also, notes still always
display their accidentals.