Honour the current key signature.
This means, first, giving DrawCtx a 'current key signature' attribute.
Make this a HasKey role, which requires a ->chroma($note) method, which
returns the default chroma for that note. Then Note->_accidental can use
that to see if an accidental should be displayed (I'm not trying for the
moment to follow the rules about accidentals carrying through the bar,
especially since I don't have any bars.)

Making the cursor follow the key signature is a little harder, at least
in insert mode. In insert mode the current cursor pitch doesn't
represent a note, but rather an octave, so the default chroma for the
about-to-be-inserted note cannot be determined until it has actually
been inserted. This means, in practice, that chroma has to join dot in
being a property that is manipulated after insertion rather than before.

Insert mode and edit mode are getting closer: length, octave and note
are the remaining properties which apply to the next rather than the
previous note in insert mode, and I suspect it would be a good idea to
change octave to always affect the note just inserted. I don't think I
can merge the two modes, though: I still need to distinguish between
'insert a C' and 'change the current note into a C'. It might actually
be clearer to make note the only property which changes behaviour:
making length something which is adjusted after insertion still keeps
the current behaviour of the next note defaulting to the same length as
the one just inserted, and (FWIW) the typed order now exactly matches
Lily: e/es/'/4/. , with chroma defaulting to follow the key signature.