Default octave to the middle of the stave.
This means a Clef method to ask what the note in the middle of the stave
is, and rather than just using centre_line it's easier to have a new
method which returns a HasPitch. There's no need for this to be a whole
Note (which HasLength as well), so create a new Util class to just
represent a bare pitch.

Move ->nearest into HasPitch, so that we are either asking $pos or
$clef->centre_pitch for the nearest N. Add a _find_ambient method since
we now need to find clefs as well as key signatures; I have Plans for
caching this information so we don't have to traverse the entire list
looking for it every time, but I haven't got to them yet.