Insert key signatures.
Add a KeySig action (bound to k) which inserts a new copy of the current
key signature. This can then be adjusted with Sharpen and Flatten. Also,
in edit mode, accept notes when on a key signature to change the key.
Take note of whether we are major or minor, and make some sensible
guesses about whether we want B or Bb major.

Move the circle-of-fifths stuff into HasKey, because the attributes want
to be Copiable, and apparently Copiable will only copy attributes from
roles. (I might want to change that...)