Zoom and scroll.
This isn't the way I want to do scroll in the end (I want to do
linewrapping) but it's getting annoying not being able to see more than
four bars at a time.

I've changed the default CTM so that we are positioned on the centre
line of the stave. I've also reduced the default zoom slightly. I
considered flipping the Y axis, since I'd really rather positive was up,
but that caused rather a mess, and I can't be bothered to sort it out...

(I'm always suspicious of overloaded operators I didn't write myself,
but I spose in this case they're OK...)