Open and save.
I can now start with an empty document.

This does not yet follow the proper SDI model: I ought to be creating a
new View (and a new Window) when I open a new Document, rather than
overwriting the current one.

The 'dirty' attribute on Document is to track unsaved files: this isn't
implemented yet.