Keep track of the ambient state.
I want to change notes so they keep a pre-rendered version of
themselves, rather than rendering a new copy every time; this is
necesssary since to keep the staffs in sync I need to know how wide
things are going to come out. This means each note needs to know,
without tracking back through the whole part, what the prevailing clef/
key/time signature is.

Each Music item now hold a ref to an Ambient node. Items which change
the ambient environment have a hard ref to their own Ambient node; these
nodes do HasAmbient. Other items have a weak ref to the Ambient
belonging to the nearest HasAmbient, and each Ambient has a weakref to
the previous Ambient. All of these refs are set up lazily, and all will
be recreated if an Ambient is destroyed.

When an Item wants to know (say) its clef, it asks the Ambient node it
points to. If it doesn't have one, one will be created, by recursing
back through the list until we get to a HasAmbient item. The Ambient
will first check to see if its owner is a clef change; if not, it will
follow its prev pointer to the previous Ambient in the list. As before,
if prev is not set, or has become undef because a weakref self-
destructed, it will be generated by recursing back along the list.