Draw notes with correct vertical alignment.
This means changing the rendering from going one voice at a time to
doing all voices simultaneously, as MIDI does. I have a new object,
BarCtx; one of these exists for each voice, and it keeps track of
barlines and temporal positioning. (I should probably superclass this to
use with MIDI, since the basic 'step through several voices in parallel'
logic is the same.)

I also need a new method on Music, ->lsb, which returns the object's
left side-bearing. This is necessary because it looks very weird if a
note on one staff is aligned with an accidental on another, rather than
with the notehead.