MIDI ties.
This means moving BarCtx to StaffCtx, with separate ::Draw and ::MIDI
subclasses for the different renderers. I also had to change the
relationship between ->when and ->item: previously ->item was the next
item to display, and ->when the time it should be displayed; this was a
little more convenient for drawing, since it meant I didn't have to
special-case the first item in a voice. Now, more straightforwardly,
->item is the currently-'playing' item, and ->when is the time remaining
until we play the next item.

Since I now have an object which wants to know which MIDI channel to
use, allocate them properly, and make Cursor get its own allocation
instead of just using channel 0.