Render to an off-screen surface.
OK, so I've given up on the idea of caching vector images.
RecordingSurface doesn't work properly, Svg- and PdfSurface have
horrible resolution-dependant artifacts, and caching as a path, while an
attractive idea, doesn't allow for alpha (which my C major key signature
needs). So, bitmaps it is (ewwww).

Set the expose event to just blit from a prerendered surface, and do the
actual rendering in the build method. View->refresh clears the surface,
so any change messages from Cursor or so still propogate properly. Since
I still can't find out bboxes without doing the real rendering, I have
to redo the whole thing if the size changes. There doesn't seem to be
any way to render to an arbitrary-sized image and trim it afterwards.

Annoyingly, this is noticeably slower when rerendering at high zoom. I
suppose the fact that the rendering is now unclipped is to blame:
before, the clipping path was set to only include the exposed part of
the image; now, every rerender renders the whole image, and just the
blit is clipped. I hate bitmaps.