Use client-side image surfaces.
Since updating the NEW_XORG, Jacana has been causing huge amounts of
memory to be Wired down, apparently not associated with any process. It
appears that create_similar_surface was creating some sort of server-
side state, which was in turn getting pushed into the kernel (...or
something?). Anyway, all I wanted was an image, so just use an
ImageSurface instead.