Double barlines.
I'm not currently trying to use Lily's \repeat mechanism, because I
don't think it wins me anything over a plain \bar. The dialog is
workable but incomplete: it has a combobox with Lily barline codes in
it, whereas it should have radio buttons with pictures. That means both
drawing the pictures and redoing Gtk's RadioButton (again...).