Refactor articulation/slurs as Marks on Notes.
There are several categories of object (articulation, slurs, dynamics)
that are essentially modifiers to a base note. Lily will only accept
them immediately following a note specification, so if they are
represented as independant Music objects there is a strong risk they
will get separated from their Note and produce invalid lily. Resolve
this by turning them into Marks, and giving Note and array of marks.

This commit breaks the editing: marks are not at this point editable.
The rendering of slurs is still messy, though not quite as bad as before
(since we now have information about the note the slur refers to). Some
of the rx/to_/from_lily logic could probably be cleaned up, especially
if I assume that marks usually take the form \foo; I may get to that
later. (I really wish (?<>) could be nested to build a tree
structure...)