Marks on rests.
I'm not entirely sure about this: obviously a staccato rest is
meaningless. However, a pause on a rest (for instance) is not, and I
believe Lily takes the easy route and just allows all marks. Certainly
something like a slur starting on a rest (which would normally be silly)
can occasionally be useful.

Possibly I want to extend Has::Marks to allow an object to specify which
types of mark are allowed. (Probably I want to separate fermata/segno/
coda, and possibly the ornaments, from Articulation, despite the fact
that Lily considers them to be the same thing.)