\mark.
Lily uses \mark for two logically rather different things: rehearsal
marks, and random text attached to a point on the stave. As far as
typesetting goes, the first is a special case of the second, but an
editor should treat them differently so I've represented them with
separate classes. Fortunately the two cases are easily distinguishable.

The generic form takes a \markup argument, so I need a role to handle
that. I've also created a Music::Text base class, since Lily has several
text-containing objects; but suddenly I'm a lot less sure it does
anything except include Has::Markup. Oh, well, I can get rid of it if I
need to...

The font scaling wasn't working properly; it turns out this is because I
had the font name wrong, so presumably FC was falling back to bitmaps or
something equally horrific.