Refactor dialogs a bit.
Dialog now takes a src attribute, which holds the object this dialog has
been invoked for. In the case of creating a new object it holds the
class name, and the copy_from is skipped. (The Text dialog will need
this, as will Tweaks.)

If I move the defaults into the dialogs themselves, I can generate all
the Cursor invoke-a-dialog methods.