Tweaks.
Allow attaching random Lily tweaks to items that support them. Currently
this is just Text::Mark, and the only tweaks supported are direction
(vertical alignment) and self-alignment-X (horizontal alignment), but
that should be easy to extend.

Create a generic dialog role, to build the tweaks section of the dialog.
This includes classes for the individual tweaks, since they need
different entry methods. (Both the tweaks that exist at the moment are
'Enum', meaning a combo box.)