Region.
I am less and less convinced a linked list is the right data structure
here. Even simple questions like 'which of these two items comes first'
are hard to answer. Hmmm.

Implement a region (selection), by giving the View an optional 'mark'
pointer. Implement octave up/down over the region, because that's easy.
At the moment cursor and mark have to be in the same voice for mark to
be useful; possibly I should move mark as cursor changes voice? Or
change mark to be a point-in-time rather than a particular note (I don't
think that would be a good idea...).