Cut and paste.
No copy yet, because I don't have a clone method on Music. I don't know
if I can get away with just copying Copiable atts; it might be enough,
though I would have to move (say) Note->octave and ->tie into a role.

Also fix a bug in LinkList->remove, which I didn't notice before because
I was only ever removing one element at a time, and then throwing it
away.

My plan was for clips to be headed by a subclass of Voice, so they could
preserve ambient key/time/&c. I haven't done that yet: the clip is just
the bare notes. Ambients generally are being annoying: I've introduced a
break_ambient method, which I'm calling everywhere; I'm not sure how
many of those calls are necessary, and I don't want to be making them at
all. I need some different system, I think; some sort of 'cached
information up to this point' object, including things like layout
position and point-in-the-bar, which can be cleared all in one go.

I would like a second form of cut which replaces the cut music with
rests, but this needs the time signature (which I can get) and the
point-in-the-bar of the start of the cut (which is harder, at the
moment). Ideally I would also want multi-measure rests (this is a really
annoying Lily feature).