Move the Region actions into a new role.
Using roles simply as a way to divide up a class seems a little odd to
me, but it makes some sense. It's necessary to load this new role after
the attributes, not before, since it requires some of the accessors.

Move the actions into their own menu, and make them apply to the whole
voice if the mark is unset, and move the mark to the current voice if
necessary.