Make LinkList generic.
I want Voices (and Movements) to be in linked lists. Since Voice isa
Music, this means Voices need to be in two different lists at the same
time. One of the advantages of linklists is that they can do this; you
just need the glue to create the attributes and methods.

Change LinkList from a role to a module which exports a new Moo keyword,
'linklist'. This takes a single parameter, the name of the list, and
creates the attributes and methods required to operate the list.

Unfortunately, although CMM provides install_modifier and RT provides
apply_roles_to_package, Moo itself provides no equivalent for 'has'.
This means that to create dynamic attributes I need an eval. Probably I
should port the whole shebang over to Moose proper at this point; I'm
sure it would make everything (including all of MooX::Misc) easier to
have a proper metaclass system. However... that's quite a bit of work.

Give Music a 'music' linklist. For now, proxy the old method names
across; I want to carry through most of the renames, but 'next' and
'prev' could usefully check for list start/end rather than being plain
accessors, and 'insert' and 'remove' are probably too Huffman to lose.