Separate out the actions/uimgr XML.
Merge the action-creating YAML into the uimgr's XML. This requires a bit
of bodging to allow me to create my radio groups. Transfer the XML to a
separate file. Read, process, and build the actions and uimgr from
BUILD, because otherwise we have to keep the parsed XML around in an
attribute forever.