Copy to clipboard.
This requires a clone method on lists. For this I am assuming all
relevant attributes are Copiable, so that means Has::Clef needs to get
its own attributes and become Copiable, and the tie att of Music::Note
needs to be moved into a Has::Tie role. (It's possible I could just make
the atts copiable in the classes, but I don't think that works
properly.)

Something rather bizarre is going on with Music::Clef->staff_line. For
some reason it's not being properly imported from Has::Clef, and instead
is inherited from Music. I can't see why this should be...