Break lines.
This is deeply crude; I make no attempt to avoid breaking in the middle
of notes, let alone bars. I also don't yet attempt to carry forward clef
and key signature. However, I really want to get incremental rendering
working, and given that I want to linebreak eventually it's simpler to
do it now. (The only sensible way I can see to do incremental rendering
renders in <some width>-chunks anyway, so those chunks might as well be
lines.)

Point-and-click is broken; I've removed View->bbox, because the same
information is (or ought to be) contained in the Lines. I just need to
rewrite the code to grovel through and find the right entry.

The vertical scroll bar (or rather, determining our total required
vertical height) is tricky. In principle the only way to find how much
height we need is to render everything; but since in principle adding or
deleting any Music item can change the display of everything else this
would make the rendering no longer incremental. So, fake it: as we
scroll down, keep telling Gtk 'just a bit more' until we get to the end.
(The 'until' part isn't working right yet.)

StaffCtx needs to acquire a clone method. Annoyingly, I can't use Clone
or any equivalent for this, because that would chase the weakrefs and
clone the whole document. So, for the moment, I've done it by hand. I
suspect I could use Copiable for this, but again I don't know whether it
works properly on classes yet.