Fix point-and-click.
I'm not sure about this solution... Replace the ->upto member of
View::Line, which was the array of StaffCtxs we had been using, with a
->staffs member which is an array of View::StaffInfo objects. These hold
enough information to recreate the StaffCtx as necessary, and also hold
bounds (temporal and physical) on the music represented by this line.

I don't like the fact that StaffInfo needs to know so much about
StaffCtx. I think perhaps I need a SystemCtx, holding an array of
vertical slices, but to do this sensibly would mean separating 'find
width' from 'draw', and I've been avoiding that because it pretty-much
doubles the work. Maybe I can make Line do the job? (Line should
probably be called System, anyway.)