Remove extraneous notes on the RHS.
Currently the note that causes the line to wrap has already been
displayed by that point. I need to change that, especially if I'm going
to try to wrap by bars, but for now just wipe out (most of) the stuff
that shouldn't be there, because it's rather confusing. Make sure not to
erase the barline, if any.