Let notes know which system they belong to.
I am inclined, at this point, to move all the bounding-box information
into the System, and just search through it when looking for a
particular note. With the bboxes in the notes, I end up searching for
point-and-click; this would mean searching for highlight, which happens
more often, but still: I suspect it's worth it.

This has to be a ref, not an index, so that it can be a weakened;
otherwise I end up having to explicitly iterate the whole document
whenever I invalidate anything. (I don't actually know the cost of
breaking a few hundred weakrefs, but iterating in C over an AV must be
cheaper than iterating in Perl over a LinkList.) This will cause
problems later, unless I turn System into a LinkList as well. I suppose
a linear search over the System array isn't too bad, given that
invalidation means I'll be doing something expensive in a minute anyway,

What I would really really like is some way of creating an array such
that objects stored in that array can find out their index cheaply, even
after splices and such. I think this ought to be possible from XS: the
array would need to be an AV-of-*Vs, like the pad arrays, so that the
real objects can be in the array rather than just refs; this means a
special constructor, since we can't just cons up a random hashref any
more. Apart from that I don't see any obvious problems...

The parameters to V::Render->_show_music are getting messier and
messier. I need to push them into some sort of SystemCtx object.