MIDI voices.
Although these are stored in the document they are not written out to
the file. I don't think the Lily syntax for setting MIDI settings is
likely to be useful for my purposes, given that I'm explicitly avoiding
making assumptions about layout and so on. Possibly I need a
Jacana-specific settings section at the head of the file, though I'd
rather avoid that if I can.

Allow <radiogroup> to contain <separator>.