Remove Span.
I'm not sure what the right thing to do about spans is, but making every
slur have its own Ambient node definitely isn't it. The whole point of
Ambients is that they are relatively rare: tracking back through all the
slurs in a piece to find the clef is not going to be much better than
simply running back through the notes.