Put MIDI instruments in the Lily.
It turns out Lily's notation is a perfectly good way to store
instruments: program changes are simply specified inline in the notes.

There is no editing yet. Has::MI->groups is for the benefit of that
dialog.

I may want to make a more general \set role, at some point.