Start Moosifying.
I've just done the easy cases first: Resource and Document. These have
nearly no is-a or does-a relationships to other classes, so they can be
converted without worrying about Moo/Moose interaction.

The exception is Has::Voices: Music::Voice is also Has::Voices, and M:V
is still a Moo class. This doesn't appear to have had any side-effects;
if this is true it will make incremental conversion much easier.