Implement copy_from.
Also move the constructor logic into a class trait (rather than an
attribute trait). This has the advantage of not having to repeat the
DOES checks for every attribute, and also means I could write a version
that works with the immutable constructor. The class trait needs to be
applied to the class the attribute finally ends up in, so applying it
with Moose::Exporter from MX:Copiable isn't any good: instead I have to
apply it when the applied_attribute gets attached to a class.

This still doesn't handle Copiable attributes in classes: I see no
particular reason not to allow this, but it would mean different logic
for finding a namespace to associate with the attributes.