Use the Attribute methods directly.
I'm not entirely sure about this: to some extent it breaks
encapsulation, as we're invoking metaattribute methods on an external
object.

I'm still not sure what to do about inlining. Unless I can find a way to
copy the copiable_atts hash down into a method on the class when an
object is made immutable, I will still end up needing to call metaclass
methods from the inlined code. And whatever I do, the inlined code can't
be that efficient, because I will have to go through the accessors for
the foreign object, at least, because I don't know which attributes need
copying at inline-time.