Deep copy.
This was a bit harder than I was expecting. Because I want to get
'required' right I need to do the copy_from processing early in the
constructor, at the point where we are gathering the init_arg list. But
I don't have access to the constructed object at that point, because it
hasn't been constructed yet, so I need to wrap a value for deep copy in
an object that can be called later on when we have an object to call it
on.