Coece is better for deep copy.
Add a Coercer att trait, for Dialog::TimeSig which needs to coerce via a
method. Add a default coerce-from-hashref for all my objects: this will
not be activated for a att without coerce=>1, but I might change that.

Put deep_copy back to just copying in a hashref and assuming coercion,
like the Moo version did.