Apply Copiable as a trait explicitly.
The 'traits' param of role attributes is a little odd: it doesn't
specify traits applied to the role att, but to the class att it will
eventually be applied as. That means I can have class and role atts both
becoming Copiable with the same syntax, but the role still needs to
import MX::Copiable to ensure the _copiable_role gets set properly.

I wonder if I should just say all atts in a Copiable role are Copiable?