Don't attempt to render notes that don't exist.
I'm not sure quite whats going on here, but _show_music is getting
called after the renderer has run off the end of the voices. I think
this happens when there is nothing to do, but the renderer is called
because Gtk has called a refresh and the music doesn't go all the way to
the botton to the page.