Fix the rerender problems.
The last line was failing to render properly every 3 keypresses. This
was because the ->continue StaffCtx object was not being cloned again,
so the second time it was used it was left at the end of the line, and
nothing got rendered. The third time, the 'current system' was
undefined, because the cursor position had never been rendered, which
made the clear_lines_from code clear all the lines and started the cycle
again.