Make the MIDI speed dialog work in bpm.
Currently the speed can only be set in crotchets-per-minute, but that
will be easy enough to fix later when I have \tempo objects.

The default timeout is now silly so it divides into 120 bpm. The bpm is
actually somewhat nominal, since Glib will truncate the timer speed to
an integer anyway. I probably need to switch to using the sequencer API.