Factor out StaffCtx into roles.
I think Cursor need to either be or have a StaffCtx, so it can handle
iterating into nested Music. (I think 'have' will be the easier choice
for now, though probably I ought to refactor all of Cursor into a
controller class.) StaffCtx atm is mostly about handling syncronisation,
which is not relevant for Cursor, so I need to make those bits optional.