Track the time at the start of each system.
Finding the current time, or finding the music at a particular time,
currently always involves iterating from the beginning. This is
necessarily slow as the movement gets longer, but I can't avoid it using
the Ambients because (a) every change would have to break the Ambient
chain and (b) it's likely that ambient nodes won't be frequent enough to
help (if we just have clef/key/time at the beginning, we still have to
iterate from there every time).

However, the rendering system ensures the rendered music is up to date,
as far as it exists, so I can store a starting tick for each system and
work from that. Mostly it is the cursor wanting to find_time (every time
the user changes voice) so mostly the music concerned should already be
rendered.