Track ticks for cursor and mark.
Make mark into a StaffCtx::Cursor (properly it wants to be a SystemCtx,
of course). Give that class Has::Tick, and keep ->tick up to date as we
move around the staff.

Currently the StaffCtx updates itself for insert and delete, but changes
to a note's length are not propogated. The next step is a time-change
signal which publishes all time changes, so anything with a tick can
make sure it stays up to date.