Start a more sensible logging system.
I want to be able to log to a logfile, and to get more useful debugging
output. However, this is slightly complicated: I want early messages to
go into the file, so we have to buffer them until we know where they're
going, but I don't want to lose them if we exit before opening a
logfile. So if we don't get that far everything gets chucked out to
STDERR as a fallback.