Log to a logfile.
Add a -d argument to request (limited) logging to STDERR instead.

This is not very good; properly I want all the logging to the logfile,
even if some goes to STDERR as well. And I need to clear up old
logfiles, and I need to remember the name of the logfile so it can be
sent in with an error report.

This all properly need to go into a separate class. Probably Log wants
to become Log::Logger, and then I have several of those, and logic for
finding and opening the logfile.

I don't know what to do about the program arguments. I was trying to
keep bin/jacana minimal, but I think perhaps all the argument handling
needs to go in there.